adjust

+++ src/migrations/20200827045151-remove-layout-setting.js
@@ -15,6 +15,10 @@ module.exports = {
 
     const layoutType = await Config.findOne({ key: 'customize:layout' });
 
+    if (layoutType == null) {
+      return;
+    }
+
     const promise = [
       // remove layout
       Config.findOneAndDelete({ key: 'customize:layout' }),
